Remove extra ajax call on page load.
If a user tries to vote while logged out, they are redirected to
the search form with the EIN pre-filled, which triggers an ajax call
to look up the charity. Instead, include the charity data as a config
value in the initial page load so we can render without an extra
request.