Add 'discuss on reddit' link.
This link appears just below the 'Vote' button on a search result, and will
do one of two things.  If the direct link to that charity has not yet been
submitted to the /r/redditdonate subreddit, it will redirect to the submit
page with the form pre-filled.  Otherwise, it will redirect to the comments
page.