Fix results showing up after deleting query.
This happened because the function that made the ajax request for a query
was debounced, but the code that cleared the results was not.  This meant that
when deleting the query character-by-character (backspacing), the last valid
query string would get processed _after_ clearing the results.  Putting the
results clearing code into the debounced function fixes it.